FIX: Minor security improvements.
Quote variables.

Instead of using a sub shell built in bash functionality is now used to strip non ascii characters from the build link generating a string that will be used as a filename.

The dawidd6/action-download-artifact is known to allow for creation of vulnerable workflows.
This workflow does not execute any code coming from artefacts. I've additionally added paths everywhere to harden against symlinks as suggested in the following link:
https://www.legitsecurity.com/blog/github-actions-that-open-the-door-to-cicd-pipeline-attacks